backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
crewai==0.51.1
crewai-tools==0.8.0
openai==1.54.3
markdown==3.7
beautifulsoup4==4.12.3
numpy==1.26.4
scikit-learn==1.5.2
tiktoken==0.7.0
